rpi3-64: Increase image size and boot partition size
We are changing things up a bit, so we're going to increase the size of
the image, and download kali-linux-default during image build, but
installation will happen during the first boot after the user logs in
and goes through the user setup.  Also increase the boot partition size
due to the increase in kernel images, just to be safe.