*: Update the scripts
All images are now ~14GB so a 16GB sdcard minimum is required.
Update the list of packages installed.
Attempt to install kali-linux-default on first boot... not sure how this is going to go, because systemd doesn't like things taking a long time. May have to move it to a script, but we'll test it out.
Add 2 new scripts, one for the NanoPi NEO Plus2 and the rpi0w that are "minimal" - no desktop system installed (basically the original scripts pre- package list updates.