rpi3-64: Various fixes and changes....
Let's try ext3 to see if this is actually the ext4 bug in the build
server's kernel.
Remove quotes from the dpkg options in smi-hack, they aren't needed,
this simplifies the service file a bit.
Add in pointing to kali-rolling before pulling in kali-linux-default so
that the service doesn't fail on first boot