veyron: More CrOS config
Based on Kapil's email to cros-dev these items are needed.

We already had DEBUG_INFO disabled.

From his email:

The relevant configuration options that need to be enabled are achieved (starting from a `.config` that is the Chrome OS config) by the steps
```
./scripts/config --enable CONFIG_VT
./scripts/config --enable CONFIG_FRAMEBUFFER_CONSOLE
./scripts/config --enable CONFIG_DRM_FBDEV_EMULATION
./scripts/config --enable CONFIG_DRM_MALI_DISPLAY
./scripts/config --enable CONFIG_ROCKCHIP_LVDS
./scripts/config --enable CONFIG_ROCKCHIP_RGB
./scripts/config --enable CONFIG_DRM_PANEL_LVDS
./scripts/config --disable CONFIG_DEBUG_INFO
make ARCH=arm olddefconfig
```
It is not clear if the last 5 options are required. Certainly the last one just removes debugging info for debugging the kernel. The other three have something to do with the panel, but one needs to experiment to see which are required.

In any case, with these configuration options, the kernel boots with virtual terminals and a usable framebuffer.