bsp: Add runonce bits
This is based on and some items pulled directly from Google's Mendel Linux.

This adds a service that lets us run scripts 1 time, typically the first boot, but
it will also check if any of the scripts have changed since being created, and if
there are changes, it will run again.

Things that will need to change:
  e2fsresize - we're going to want to expand the partition first
  enable-services - decide a list of services we're going to want to enable - possibly do services based on files in /boot (ala enable-ssh)

  The user and language related stuff, we can probably make it so that they can be replaced with a supplied user variable in the build script or something,
  and replace that, so when users create their own images, they can generate them with their own users/language settings.  For now, we hardcode en_US.UTF-8 and kali