common: Fix ups
Passing $extra_args in quotes causes it to try to pass the full extra
args as the hostname, so, don't do quote it (and add a comment so that
others who come after won't do it either).

Swap the logic in the systemd_nspawn loop, so we check for the highest
version first.