common: actually do something with the mirror and suite
We have the ability to use a custom mirror and suite, so lets actually do something with them.

This is a bit of a hack really, and doesn't take into account if someone only specifies a custom mirror and not suite, I haven't tested what happens in that case, and it might break, but we will see.