neoplus2: Clean up bluetooth and wifi a bit
This removes all the old bluetooth bits that were for the sunxi kernel.  We use the mainline kernel now, so we can use the firmware-brcm80211 package, however we do still need a file from the github repo since it's not in the upstream firmware repository.

Also put the 2 bluetooth firmware files into /lib/firmware/brcm instead of ap6212, since we aren't using the old bluetooth way.

Bluetooth still doesn't work, but wifi does.