regenerate ssh host keys: remove un-needed entries
These don't belong there, and aren't needed.  It was an attempt to make sure it was run before ssh starts, but making it part of when put in the correct spot in the service file seems to cause ssh to just never actually start, so lets remove it.  We can revisit if it ever causes issues.