pinebook-pro: Add 2GHz OPP, protected by turbo-boost
This adds 2GHz as a possible CPU frequency, but it's protected behind turbo-boost

In order to enable, you will need to

echo 1 > /sys/devices/system/cpu/cpufreq/boost

which is easiest done by adding it to rc.local if you want to use it.

It is disabled by default, but can also be disabled by simply echoing 0 to the same file.