pinebook: Fixups
The 5.14 kernels make warnings errors, and when a driver has the date and time built into it, that makes it a reproducable build warning, which is now an error, so sed that out.

Also, it seems as though u-boot-menu no longer runs u-boot-update automatically, so we run it to generate the /boot/extlinux/extlinux.conf file even though it will be empty, so we can use it later.