zero-2-w: Add new device script and firmwares
Currently the kernel packages don't include the dtbs and the firmware package doesn't have the firmwares, so add them in for now as a workaround.  Also point to re4son's next repo for the latest kernel that has the changes needed for the dtbs to work.

Note: the current boot firmware expects the dtb name to be bcm2710-rpi-zero-2 which is why we include both, so that when the new bootloader shows up, it will seamlessly transition to the correct one.