usbarmory*: Update u-boot, fix up u-boot builds and update kernel compiles.
The usbarmory mk ii u-boot wasn't actually building for the usbarmory mkii, so we grab the patch from the git repo and enable usbarmory mk ii build support.

Add in patch for usbarmory mk1's microsd card detection in u-boot.

Update the github repo to point to the new repos instead of relying on github to redirect us.

Remove a removed dts file upstream for the usbarmory mkii.

Bump the u-boot used by both to 2021.07.