usbarmory: Fix the builds
The kernel was bumped to 5.15, so we patch the 5.15 wifi injection in.
Also, the defconfig file needs to be named _defconfig for copying it in and using it as a target.
With newer u-boot, we don't seem to need the patch for usbarmory mki
Patch from upstream for mkii u-boot requires newer u-boot as well, so bump it.