Put the kali user into a file in /etc/sudoers.d
When running updates, having our kali user in /etc/sudoers itself,
causes an update to ask if the user wants to use upstream or not.  Lets
add them to their own file in /etc/sudoers.d so that it doesn't cause an
update to ask about the changes, and keeps our change to not require
entering the password when using sudo on arm devices.